Use unsigned ints where they are needed

+++ src/intel.c
@@ -167,7 +167,7 @@ static intel_uarch_t brand_string_method(const cpuid_data_t *data)
     intel_uarch_t uarch = NO_CODE;
     const char *bs = data->brand_str;
     char *p = NULL;
-    int i, need_table = 1;
+    unsigned int i, need_table = 1;
 
     const struct {
         intel_uarch_t uarch;
@@ -241,7 +241,7 @@ static intel_uarch_t brand_string_method(const cpuid_data_t *data)
             uarch = MOBILE_PENTIUM;
     }
     if (need_table) {
-        for (i = 0; i < (int)NELEMS(uarch_t); i++)
+        for (i = 0; i < NELEMS(uarch_t); i++)
             if (match_pattern(bs, uarch_t[i].name)) {
                 uarch = uarch_t[i].uarch;
                 break;
